This commit was manufactured by cvs2svn to create branch 'GlastReleasev9r24'.
Sprout from master 2007-01-03 03:56:05 UTC richard 'add microQuasar'
Cherrypick from master 2005-03-06 22:24:44 UTC jchiang 'AutoUpdate':
    genericSources/ChangeLog
    genericSources/cmt/requirements
    genericSources/doc/release.notes
    genericSources/genericSources/EblAtten.h
    genericSources/genericSources/MapCube.h
    genericSources/genericSources/MapSource.h
    genericSources/genericSources/SpectralTransient.h
    genericSources/src/EblAtten.cxx
    genericSources/src/FitsImage.cxx
    genericSources/src/FitsImage.h
    genericSources/src/IRB_routines.cxx
    genericSources/src/MapCube.cxx
    genericSources/src/MapSource.cxx
    genericSources/src/PeriodicSource.cxx
    genericSources/src/Pulsar.cxx
    genericSources/src/SimpleTransient.cxx
    genericSources/src/SpectralTransient.cxx
    genericSources/src/TransientTemplate.cxx
    genericSources/src/Util.cxx
    genericSources/src/Util.h
    genericSources/src/mainpage.h
    genericSources/src/sources.in
    genericSources/src/test/test.cxx
    genericSources/xml/source_library.xml
Cherrypick from master 2007-01-30 00:14:29 UTC richard 'AutoUpdate':
    genericSources/genericSources/TF1Spectrum.h
    genericSources/src/TF1Map.cxx
    genericSources/src/TF1Spectrum.cxx
    microQuasar/ChangeLog
    microQuasar/microQuasar/microQuasar.h
    microQuasar/src/microQuasar.cxx
Cherrypick from master 2006-11-06 10:50:00 UTC omodei 'AutoUpdate':
    GRB/ChangeLog
    GRBtemplate/ChangeLog
    Pulsar/ChangeLog
Cherrypick from master 2005-11-09 00:32:59 UTC lreyes 'AutoUpdate':
    eblAtten/ChangeLog
Cherrypick from master 2006-03-21 18:59:27 UTC usher 'Updates for CLHEP 1.9.2.2':
    Pulsar/src/PulsarSpectrum.cxx
Cherrypick from master 2006-10-31 22:11:40 UTC omodei 'AutoUpdate':
    ChangeLog
Cherrypick from master 2006-11-01 16:36:13 UTC omodei 'AutoUpdate':
    SpectObj/ChangeLog
Cherrypick from master 2006-11-08 11:12:28 UTC omodei 'AutoUpdate':
    GRBobs/ChangeLog
Cherrypick from GLAST 2005-08-29 21:23:09 UTC jchiang 'initial import':
    .cvsignore
    GRBobs/.cvsignore
    GRBobs/src/test/other/compile.C
    GRBtemplate/.cvsignore
    eblAtten/.cvsignore
    genericSources/.cvsignore
    genericSources/data/3c279Template.dat
    genericSources/data/CrabTemplate.dat
    genericSources/data/GemingaTemplate.dat
    genericSources/genericSources/GaussianSource.h
    genericSources/genericSources/Isotropic.h
    genericSources/genericSources/PeriodicSource.h
    genericSources/genericSources/Pulsar.h
    genericSources/genericSources/SimpleTransient.h
    genericSources/src/GaussianSource.cxx
    genericSources/src/Isotropic.cxx
    src/mainpage.h
Delete:
    DarkMatter/.cvsignore
    DarkMatter/ChangeLog
    DarkMatter/DarkMatter/Profile.h
    DarkMatter/cmt/.cvsignore
    DarkMatter/cmt/requirements
    DarkMatter/release.notes
    DarkMatter/sources/DS_DC2/liste.dat
    DarkMatter/sources/DS_DC2/param.dat
    DarkMatter/sources/DS_DC2/res100MeVBisBis.txt
    DarkMatter/sources/DS_DC2/res10MeVBisBis.txt
    DarkMatter/sources/DS_DC2/rndm00000060.dat
    DarkMatter/sources/DS_DC2/rndm00000080.dat
    DarkMatter/sources/DS_DC2/rndm00000100.dat
    DarkMatter/sources/DS_DC2/rndm00000120.dat
    DarkMatter/sources/DS_DC2/rndm00000140.dat
    DarkMatter/sources/DS_DC2/rndm00000160.dat
    DarkMatter/sources/DS_DC2/rndm00000180.dat
    DarkMatter/sources/Profiles/1e10-85kpc_map.fits
    DarkMatter/src/LineOfSight.cxx
    DarkMatter/src/LineOfSight.h
    DarkMatter/src/Profile.cxx
    DarkMatter/src/globals.h
    DarkMatter/src/haloProfile.cxx
    DarkMatter/src/mainpage.h
    DarkMatter/src/makeMap.cxx
    DarkMatter/xml/darkmatter_library.xml
    pySources/.cvsignore
    pySources/ChangeLog
    pySources/cmt/requirements
    pySources/doc/release.notes
    pySources/python/ebl_plot.py
    pySources/python/test_sources.py
    pySources/src/pySources.in

+++ microQuasar/microQuasar/microQuasar.h
@@ -155,7 +155,7 @@ private:
 
 	/// fiddle of Numerical Recipe's rtsafe to avoid function passing
 	double rtsafe(const double x1, const double x2, const double xacc);
-	std::pair<float,float> calculateJetStart(bool nextOn, float time);
+	std::pair<double,double> calculateJetStart(bool nextOn, double time);
 
 	/// flux (ph s^-1 cm^-2)
 	float m_ftot;
@@ -176,8 +176,8 @@ private:
 	float m_randPhase;	
 	/// current time
 	double m_currentTime;
-	float m_jetStart;
-	float m_jetEnd;
+	double m_jetStart;
+	double m_jetEnd;
 	/// disk-cycle properties
 	DiskCycleProperties m_diskProperties;
 	/// jet timing properties
